Update product.php
In the PHP code, $product['filename'] is checked to see if it is empty. If it is not empty (that is, an image exists), use the HTML img tag to display the image. The img tag will only be output if $product['filename'] is not empty. This means that img tags will not be generated for pages that have no images or have removed images.