Create search.php
A search form is available at the top of all pages. The keyword or keywords entered into the search form will be used to search for pages that include the provided word or phrase. At a minimum the page name will be searched using a SQL LIKE query with wildcards, but other page properties can also be searched. The search will result in a list of links to all found pages.